Refactor: move helper modules to helpers/ folder
