+++ kustomize/commands/__init__.py
@@ -1,0 +1,3 @@
+"""
+CLI command modules
+"""
