+++ tests/test_cli.py
@@ -4,7 +4,7 @@ Tests for command line interface (CLI)
 import os
 import pytest
 
-import kustomize.cli
+import kustomize
 
 
 def test_runas_module():
