+++ tests/test_commands_lint.py
