+++ tests/test_helpers_binaries.py
@@ -6,7 +6,7 @@ import sys
 
 from unittest.mock import patch
 
-import kustomize.binaries
+import kustomize
 
 
 @patch('platform.system', return_value='Linux')
@@ -17,34 +17,34 @@ def test_binarypath(mock_platform):
     old_prefix = sys.prefix
     sys.prefix = '/some/path'
 
-    path = kustomize.binaries.binarypath('foo')
+    path = kustomize.helpers.binaries.binarypath('foo')
     assert str(path) == '/some/path/local/bin/foo'
 
     sys.prefix = old_prefix
 
 
-@patch('kustomize.binaries.run')
+@patch('kustomize.helpers.binaries.run')
 def test_run_piped_commands(mock_run):
     """
     Is list of commands properly executed as a pipe?
     """
     command_list = ['foo abc', 'bar -v', 'baz']
-    kustomize.binaries.run_piped_commands(command_list)
+    kustomize.helpers.binaries.run_piped_commands(command_list)
 
     assert len(mock_run.mock_calls) == 3, \
         f"run() not called for each command in '{' | '.join(command_list)}'"
 
 
 @patch('builtins.print')
-@patch('kustomize.binaries.run_piped_commands')
+@patch('kustomize.helpers.binaries.run_piped_commands')
 def test_shell_command(mock_run_piped_commands, mock_print):
     """
     Is command printed and then executed?
     """
-    executable = kustomize.binaries.binarypath('foo')
+    executable = kustomize.helpers.binaries.binarypath('foo')
     exec_location = str(executable.parent) + os.path.sep
     shell_command = f"{executable} --bar | {executable} baz"
-    kustomize.binaries.shell(shell_command)
+    kustomize.helpers.binaries.shell(shell_command)
 
     assert mock_print.called, \
         "print() is never called"
@@ -58,7 +58,7 @@ def test_shell_failing_returncode():
     """
     Does an invalid command return a failing status code?
     """
-    result = kustomize.binaries.shell('/non/existing/command')
+    result = kustomize.helpers.binaries.shell('/non/existing/command')
 
     assert result.returncode, \
         "Non-zero status code expected, zero received"
