+++ tox.ini
@@ -55,7 +55,7 @@ exclude = .tox,build,dist,tests
 [flake8]
 exclude = .tox,build,dist,kustomize.egg-info
 per-file-ignores =
-   kustomize/download.py:E501
+   kustomize/helpers/download.py:E501
 
 [pytest]
 addopts =
